refactor internals into a module
